basic sanity kits for stage-a
