+++ test/manual/dsv4/test_dsv4_flash_sanity_dp4.py
@@ -3,7 +3,9 @@
 import unittest
 
 from sglang.srt.utils import kill_process_tree
-from sglang.test.kits.server_sanity_kit import ServerSanityMixin
+from sglang.test.kits.basic_api_contract_kit import BasicAPIContractMixin
+from sglang.test.kits.basic_decode_correctness_kit import BasicDecodeCorrectnessMixin
+from sglang.test.kits.basic_scheduler_stress_kit import BasicSchedulerStressMixin
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
@@ -46,7 +48,12 @@ _EAGLE_SPEC_ARGS = [
 ]
 
 
-class TestDSV4FlashTP4DP4(ServerSanityMixin, CustomTestCase):
+class TestDSV4FlashTP4DP4(
+    BasicAPIContractMixin,
+    BasicDecodeCorrectnessMixin,
+    BasicSchedulerStressMixin,
+    CustomTestCase,
+):
     """TP4 + DP4 + deepep + EAGLE MTP."""
 
     @classmethod
@@ -79,7 +86,12 @@ class TestDSV4FlashTP4DP4(ServerSanityMixin, CustomTestCase):
         kill_process_tree(cls.process.pid)
 
 
-class TestDSV4FlashTP4EP(ServerSanityMixin, CustomTestCase):
+class TestDSV4FlashTP4EP(
+    BasicAPIContractMixin,
+    BasicDecodeCorrectnessMixin,
+    BasicSchedulerStressMixin,
+    CustomTestCase,
+):
     """TP attn + EP MoE (no DP attn) — exercises the DeepEP + TP-attn path."""
 
     @classmethod
@@ -112,7 +124,12 @@ class TestDSV4FlashTP4EP(ServerSanityMixin, CustomTestCase):
         kill_process_tree(cls.process.pid)
 
 
-class TestDSV4FlashTP4DP4ChunkedPrefillLarge(ServerSanityMixin, CustomTestCase):
+class TestDSV4FlashTP4DP4ChunkedPrefillLarge(
+    BasicAPIContractMixin,
+    BasicDecodeCorrectnessMixin,
+    BasicSchedulerStressMixin,
+    CustomTestCase,
+):
     """TP4 + DP4 with --chunked-prefill-size 16384 — large chunked prefill."""
 
     @classmethod
