+++ test/registered/dsv4/test_deepseek_v4_flash_fp4_megamoe_b200.py
@@ -12,7 +12,9 @@ from types import SimpleNamespace
 
 from sglang.srt.utils import kill_process_tree
 from sglang.test.ci.ci_register import register_cuda_ci
-from sglang.test.kits.server_sanity_kit import ServerSanityMixin
+from sglang.test.kits.basic_api_contract_kit import BasicAPIContractMixin
+from sglang.test.kits.basic_decode_correctness_kit import BasicDecodeCorrectnessMixin
+from sglang.test.kits.basic_scheduler_stress_kit import BasicSchedulerStressMixin
 from sglang.test.run_eval import run_eval
 from sglang.test.test_utils import (
     DEFAULT_URL_FOR_TEST,
@@ -54,7 +56,12 @@ def _gsm8k_check(test_case):
     test_case.assertGreater(metrics["score"], 0.93)
 
 
-class TestDSV4FlashFP4B200W4A8MegaMoE(ServerSanityMixin, CustomTestCase):
+class TestDSV4FlashFP4B200W4A8MegaMoE(
+    BasicAPIContractMixin,
+    BasicDecodeCorrectnessMixin,
+    BasicSchedulerStressMixin,
+    CustomTestCase,
+):
     """Balanced recipe: TP=4, DP=4, MegaMoE."""
 
     @classmethod
@@ -95,7 +102,12 @@ class TestDSV4FlashFP4B200W4A8MegaMoE(ServerSanityMixin, CustomTestCase):
         _gsm8k_check(self)
 
 
-class TestDSV4FlashFP4B200W4A4MegaMoE(ServerSanityMixin, CustomTestCase):
+class TestDSV4FlashFP4B200W4A4MegaMoE(
+    BasicAPIContractMixin,
+    BasicDecodeCorrectnessMixin,
+    BasicSchedulerStressMixin,
+    CustomTestCase,
+):
     """Balanced recipe: TP=4, DP=4, MegaMoE."""
 
     @classmethod
